RF: move DueCredit dataset treatment to .support.due_utils
